code/requirements.txt
fastapi
uvicorn
pypdf
asyncpg
pgvector
ollama
python-dotenv
multipart
semantic-text-splitter
